Split rs.py into multiple files
